Fix compiling with clang-cl inside a Visual Studio 2019 16.4 command prompt.
This was introduced by 8586e62c7c9360327a490e67d7feaf77be01d4ee and was preventing from compiling with clang-cl on Windows.
The problem was that clang-cl detects the triple from the current env vars (was x86_64-pc-windows-msvc19.24.28315 for me, as I happen to always run inside a VS2019 cmd prompt).