Revert "[DebugInfo][DAG] Distinguish different kinds of location indirection"
This reverts commit c87c9e20708af57a68a79900caa88ff096581447.

I'm backing out D68945, which this patch is a follow up for. It'll be
re-landed when D68945 is fixed.

The changes to dbg-value-func-arg.ll occur because our handling of certain
kinds of location now mixes up indirection that happens at different points
in a DIExpression. While this is a regression, it's a return to the prior
behaviour while a better patch is sought.