[AArch64][GlobalISel] Reland SLT/SGT TBNZ optimization
The issue in the previous commits was that we swap the LHS and RHS while
looking for the constant. In SLT/SGT, the constant must be on the RHS, or the
optimization is invalid.

Move the swapping logic after the check for the SLT/SGT case and update tests.

Original commits:

8a1cb51b44194148fdf8cb344873f8c1fc053c93
e5ddca0eafac5a1edd47287b5116b4d46488ac1c