Reapply "[OpenMP][IRBuilder] Perform finalization (incl. outlining) late"
Reapply 64b5457f8d98a20167e9ecbd84c22dcf0c326665 with minor fixes.

The problem was that cancellation can cause new edges to the parallel
region exit block which is not outlined. The CodeExtractor will encode
the information which "exit" was taken as a return value. The fix is to
ensure we do not return any value from the outlined function, to prevent
control to value conversion we ensure a single exit block for the
outlined region.

This reverts commit d8d26a851001b8307bb352791e6021f357149e1a.